requirements.txt
# Google APIs Client Library
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0

# Environment variables
python-dotenv==1.0.0
